+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  - Segments straddle arc in linearization.
  - Control max-arc-error with GCode var.
  - Implemented path modes G61, G61.1 & G64 with naive CAM but not blending, yet.
+ - Handle LASER intensity correctly for M4 as well as M3.
 
 ## v1.2.0:
  - Improved Buildbotics CNC controller connectivity.